move syntax defs that start with #', ##' before rOBlock*, rOTitle*, due to :h:syn-priority
These include rOCommentKey and rOR6Block.

The lookbehind regex (\@<=) made it possible for rOBLock*, rOTitle*
and rOCommentKey, rOR6Block, to begin at the same location.
This results in, say, rOCommentKey masking the syntax of an rOBlock.
Moving syntax definitions that start with #', ##' but are not part
of an rOBlock* or rOTitle* resolves this issue